رفع مشکل جزییات صفحه سرمایه گذاری من در صورت بزرگ بودن عدد سرمایه گذازی شده

+++ HamAfarin/Classes/JsonModel/FaraboorsResponseJsonModel.cs
@@ -3,7 +3,7 @@
     public class FaraboorsResponseJsonModel
     {
         public string TraceCode { get; set; }
-        public int ProvidedFinancePrice { get; set; }
+        public long ProvidedFinancePrice { get; set; }
         public string Message { get; set; }
     }
 }
